AG-13668 Move PickFocus* types into their own file
