+++ packages/ag-charts-community/src/chart/series/cartesian/bubbleSeries.ts
@@ -27,8 +27,8 @@ import type { LegendSymbolOptions } from '../../legend/legendSymbol';
 import type { Marker } from '../../marker/marker';
 import { getMarker } from '../../marker/util';
 import { type TooltipContent, type TooltipContentDataRow } from '../../tooltip/tooltip';
-import type { PickFocusInputs, SeriesNodeEventTypes } from '../series';
-import { SeriesNodePickMode } from '../series';
+import type { PickFocusInputs } from '../pickFocus';
+import { type SeriesNodeEventTypes, SeriesNodePickMode } from '../series';
 import { resetLabelFn, seriesLabelFadeInAnimation } from '../seriesLabelUtil';
 import { type BubbleNodeDatum, BubbleSeriesProperties } from './bubbleSeriesProperties';
 import type { CartesianAnimationData } from './cartesianSeries';
