+++ packages/ag-charts-community/src/chart/series/cartesian/lineSeries.ts
@@ -33,7 +33,8 @@ import { type LegendSymbolOptions } from '../../legend/legendSymbol';
 import type { Marker } from '../../marker/marker';
 import { getMarker } from '../../marker/util';
 import { type TooltipContent } from '../../tooltip/tooltip';
-import { type PickFocusInputs, SeriesNodePickMode } from '../series';
+import { type PickFocusInputs } from '../pickFocus';
+import { SeriesNodePickMode } from '../series';
 import { resetLabelFn, seriesLabelFadeInAnimation } from '../seriesLabelUtil';
 import { datumStylerProperties, visibleRangeIndices } from '../util';
 import type { CartesianAnimationData } from './cartesianSeries';
