+++ packages/ag-charts-community/src/chart/series/dataModelSeries.ts
@@ -6,7 +6,8 @@ import { clamp } from '../../util/number';
 import { ChartAxisDirection } from '../chartAxisDirection';
 import type { DataController } from '../data/dataController';
 import type { DataModel, DataModelOptions, ProcessedData, PropertyDefinition } from '../data/dataModel';
-import type { PickFocusInputs, PickFocusOutputs, SeriesNodeDataContext } from './series';
+import type { PickFocusInputs, PickFocusOutputs } from './pickFocus';
+import type { SeriesNodeDataContext } from './series';
 import { Series } from './series';
 import type { SeriesProperties } from './seriesProperties';
 import type { SeriesNodeDatum } from './seriesTypes';
